drop -1 workaround for enum CURLMSG

+++ tests/libtest/lib530.c
@@ -203,8 +203,8 @@ static int t530_checkForCompletion(CURLM *multi, int *success)
         *success = 0;
     }
     else {
-      curl_mfprintf(stderr, "%s got an unexpected message from curl: %i\n",
-                    t530_tag(), message->msg);
+      curl_mfprintf(stderr, "%s got an unexpected message from curl: %d\n",
+                    t530_tag(), (int)message->msg);
       result = 1;
       *success = 0;
     }
